Refactored morgan into middleware folder
