+++ server.js
@@ -30,8 +30,8 @@ const dbParams = require('./lib/db.js');
 const usersRoutes = require("./routes/users");
 const widgetsRoutes = require("./routes/widgets");
 
-// ?? Utilities
-const morgan = require('./utils/morganCustom');
+// ?? Middleware
+const morgan = require('./middleware/morganCustom');
 
 // ???????????????????????? Node Modules ??????????????????????????
 const path = require('path');
